fix(arch): move wait_for_signal to nexus.lib to fix tier boundary
system_services cannot import from bricks (five-tier architecture).
Move wait_for_signal() from nexus.bricks.ipc.wakeup to nexus.lib.pipe_wakeup.
Re-export from bricks.ipc.wakeup for backward compat.
